Tests for port security and allowed address pairs
Added helper functions for packet capture and sending to VTM Guest.

Add cleanup_vms function to NeutronTestCase

Add ICMP rule to PCAP_Filter for icmp protocol

Add check in CLI.cmd() call so that the return value is stored even
  if return output is true.  Also check this value in Guest.execute()
  and throw if it is not 0.

Add test for allowed_address_pairs

Add test for port_security

Update tests to use cleanup utility function to clean up VMs and ports

+++ common/TCPDump.py
@@ -175,6 +175,9 @@ class TCPDump(object):
         self.tcpdump_stop.set()
 
         # Join the thread and if it hasn't finished already, kill it
+        if self.process is None:
+            return None
+
         self.process.join(5)
         if self.process.is_alive():
             self.process.terminate()
@@ -212,7 +215,7 @@ class TCPDump(object):
                       "'" + (pcap_filter.to_str() if pcap_filter is not None else '') + "'" + \
                       ' >> ' + tmp_dump_filename
 
-            #cli.log_cmd = True
+            cli.log_cmd = True
 
             # FLAG STATE: ready[clear], stop[clear], finished[clear]
